Fix "You're compiling with static library settings" warning when including ComponentLoader.hpp from non-component libraries
The warning is only supposed to be emitted when rtt/Component.hpp is included outside of component libraries for non-static builds.
The solution is to move the definitions of some typedefs and helper classes to ComponentLoader.hpp and do not include Component.hpp from there.
